Refactor: Updated the production code for handeling the perimeter of canvas to handle the decimal values

+++ src/calculateCanvasPerimeter/calculateCanvasPerimeter.ts
@@ -2,5 +2,5 @@ export default function calculateCanvasPerimeter(
   length: string,
   width: string
 ) {
-  return 2 * (parseInt(length) + parseInt(width));
+  return 2 * (parseFloat(length) + parseFloat(width));
 }
